requirements.txt
streamlit==1.34.0
PyPDF2==3.0.1
python-dotenv==1.0.1
google-generativeai==0.4.1
pandas==2.2.2 